+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Dev
+
+- Correctly pass generic method type and lifetime arguments to the delegated method.
+
 # 0.13.0 (2. 9. 2024)
 
 - Generalize match arms handling. You can now combine a match expression target with annotations like `#[into]` and